Create Actual Modifications
This is meant to be inserted starting at line 933 of formats.js. The old code that was at lines 933-938 will have to be deleted. Changes inlcude: more bans for pokemon that cannot be in C&E at all, a ban for the uber-tier mega stones, removal of the useless "Standard Pokebank" Ruleset (No longer in the rest of the code due to updates on the main server, it does not actually do anything), and the implementation of the following clauses: Sleep Clause, Moody Clause, OHKO Clause, HP Percentage Mod, and Evasion Moves Clause. We hope to implement a minimum level of 100 but are still working on how to do so safely. Note that this does not include brackets {}, or replace them, so please simply replace the code between the brackets at 932 and 939 with this.